requirements.txt
absl-py==2.4.0
altair==6.0.0
annotated-doc==0.0.4
annotated-types==0.7.0
ansicon==1.89.0
anyio==4.12.1
asgiref==3.11.1
asttokens==3.0.1
astunparse==1.6.3
attrs==25.4.0
bcrypt==5.0.0
blessed==1.33.0
blinker==1.9.0
cachetools==7.0.2
catalogue==2.0.10
certifi==2026.2.25
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
contexttimer==0.3.3
contourpy==1.3.3
cycler==0.12.1
decorator==5.2.1
distlib==0.4.0
Django==6.0.3
djangorestframework==3.16.1
executing==2.2.1
fastapi==0.135.1
filelock==3.20.3
Flask==3.1.2
Flask-JWT-Extended==4.7.1
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
flatbuffers==25.12.19
fonttools==4.61.1
fsspec==2026.2.0
gast==0.7.0
gitdb==4.0.12
GitPython==3.1.46
google-pasta==0.2.0
greenlet==3.3.1
grpcio==1.78.0
gunicorn==25.0.1
h11==0.16.0
h5py==3.15.1
humanfriendly==10.0
idna==3.11
ipython==9.11.0
ipython_pygments_lexers==1.1.1
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
jinxed==1.3.0
joblib==1.5.3
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
keras==3.13.2
kiwisolver==1.4.9
libclang==18.1.1
Markdown==3.10.2
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.10.8
matplotlib-inline==0.2.1
mdurl==0.1.2
ml_dtypes==0.5.4
mpmath==1.3.0
namex==0.1.0
narwhals==2.17.0
networkx==3.6.1
nltk==3.9.2
npm==0.1.1
numpy==2.4.2
opt_einsum==3.4.0
optional-django==0.1.0
optree==0.19.0
packaging==26.0
pandas==2.3.3
parso==0.8.6
pillow==12.1.1
platformdirs==4.5.1
prompt_toolkit==3.0.52
protobuf==6.33.5
psycopg2-binary==2.9.11
pure_eval==0.2.3
pyarrow==23.0.1
pydantic==2.12.5
pydantic_core==2.41.5
pydeck==0.9.1
Pygments==2.19.2
PyJWT==2.11.0
pyparsing==3.3.2
pyreadline3==3.5.4
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2026.1.post1
referencing==0.37.0
regex==2026.1.15
requests==2.32.5
rich==14.3.3
rpds-py==0.30.0
scikit-learn==1.8.0
scipy==1.17.0
seaborn==0.13.2
setuptools==82.0.0
six==1.17.0
smmap==5.0.2
SQLAlchemy==2.0.48
sqlparse==0.5.5
srsly==2.5.2
stack-data==0.6.3
starlette==0.52.1
streamlit==1.55.0
streamlit-drawable-canvas==0.9.3
sympy==1.14.0
tenacity==9.1.4
tensorboard==2.20.0
tensorboard-data-server==0.7.2
tensorflow==2.20.0
termcolor==3.3.0
textblob==0.19.0
threadpoolctl==3.6.0
toml==0.10.2
torch==2.10.0
torchvision==0.25.0
tornado==6.5.4
tqdm==4.67.2
traitlets==5.14.3
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
unicorn==2.1.4
urllib3==2.6.3
uvicorn==0.41.0
virtualenv==20.36.1
watchdog==6.0.0
wcwidth==0.6.0
Werkzeug==3.1.5
wheel==0.46.3
whitenoise==6.12.0
wrapt==2.1.1
